+++ internal/models/partner_profile/partner_profile.go
@@ -2,8 +2,8 @@ package partner_profile
 
 // PartnerProfile represents a partner's profile information
 type PartnerProfile struct {
-	ID        int    `db:"id" json:"id"`
-	AccountID int    `db:"account_id" json:"account_id"`
+	ID        int64  `db:"id" json:"id"`
+	AccountID int64  `db:"account_id" json:"account_id"`
 	ShopName  string `db:"shop_name" json:"shop_name"`
 	PrinterID string `db:"printer_id" json:"printer_id,omitempty"`
 }
